requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
google-genai
bcrypt
python-jose[cryptography]
email-validator